requirements.txt
aniso8601==7.0.0
asgiref==3.3.4
autopep8==1.5.6
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-filter==2.4.0
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
flake8==3.9.1
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==2.10
importlib-metadata==4.0.1
mccabe==0.6.1
plaid-python==7.3.0
promise==2.3
psycopg2==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==1.7.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
Rx==1.6.1
singledispatch==3.6.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
zipp==3.4.1
